Revert "Built production"
This reverts commit a0b1d94be1d8157fe354b2c185d1ebe0535c2034.

+++ deployment.sh
@@ -1,6 +1,3 @@
-if [ -d "./docs"]; then
-     rm -rf ./docs && rmdir ./docs
-fi
-
+test -e ./docs && rm -rf ./docs && rmdir ./docs
 npm run build
 mv ./build ./docs
